refactor: atoms dependency on web through AdditionalCalendarSelector
